--- src/main/java/kr/hhplus/be/server/product/infrastructure/ProductRepository.java
@@ -1,7 +1,0 @@
-package kr.hhplus.be.server.product.infrastructure;
-
-import kr.hhplus.be.server.product.domain.Product;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface ProductRepository extends JpaRepository<Product, Long> {
-}
