+++ src/main/java/kr/hhplus/be/server/product/infrastructure/repository/ProductJpaRepository.java
@@ -1,0 +1,7 @@
+package kr.hhplus.be.server.product.infrastructure.repository;
+
+import kr.hhplus.be.server.product.domain.Product;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface ProductJpaRepository extends JpaRepository<Product, Long> {
+}
